Modified pdf view and generation
